+++ docs/releases.md
@@ -38,6 +38,7 @@ changes over time.
 * github-444: Remove Flowman Hub
 * github-443: Remove Flowman DSL
 * github-466: Use frontend-maven-plugin for all npm packages
+* github-467: Implement "target_records" metric for "copy" target
 
 
 ### Version 1.1.0 - 2023-10-10